Update godot_cpp to latest 4.4 branch changes, create xml documentation using Godot's doc tool, fix some code issues and add new BlastBullets2D logo

+++ main/src/bullets/block_bullets2d.hpp
@@ -41,7 +41,7 @@ public:
 
         Vector2 cache_velocity_calc = all_cached_velocity[0] * delta;
 
-        for (size_t i = 0; i < amount_bullets; i++) {
+        for (int i = 0; i < amount_bullets; i++) {
             if (bullets_enabled_status[i] == false) {
                 continue;
             }
